Database_Connection/RestAPI_mysql/requirements.txt
Flask
mysql-connector-python
requests
pytest